+++ NEWS.md
@@ -1,3 +1,8 @@
+# 0.2 (2013-12-13)
+
+* Support RDF containers (rdf:Seq, rdf:Bag, rdf:Alt).
+* Support multiple creators per work.
+* Provide semantic markup in formatters which support property semantics.
 
 # 0.1 (2013-11-18)